aksh-api/requirements.txt
Django<3.3
django-filter<2.5
djangorestframework<3.13
psycopg2<2.9
